+++ src/js/components/htmleditor.js
@@ -25,6 +25,7 @@
             mode         : 'split',
             markdown     : false,
             autocomplete : true,
+            enablescripts: false,
             height       : 500,
             maxsplitsize : 1000,
             codemirror   : { mode: 'htmlmixed', lineWrapping: true, dragDrop: false, autoCloseTags: true, matchTags: true, autoCloseBrackets: true, matchBrackets: true, indentUnit: 4, indentWithTabs: false, tabSize: 4, hintOptions: {completionSingle:false} },
@@ -289,6 +290,10 @@
 
             this.currentvalue = this.editor.getValue();
 
+            if (!this.options.enablescripts) {
+                this.currentvalue = this.currentvalue.replace(/<(script|style)\b[^<]*(?:(?!<\/(script|style)>)<[^<]*)*<\/(script|style)>/img, '');
+            }
+
             // empty code
             if (!this.currentvalue) {
 
@@ -479,11 +484,44 @@
             });
 
             editor.htmleditor.on('click', 'a[data-htmleditor-button="fullscreen"]', function() {
+
                 editor.htmleditor.toggleClass('uk-htmleditor-fullscreen');
 
                 var wrap = editor.editor.getWrapperElement();
 
                 if (editor.htmleditor.hasClass('uk-htmleditor-fullscreen')) {
+
+                    var fixedParent = false, parents = editor.htmleditor.parents().each(function(){
+                        if (UI.$(this).css('position')=='fixed' && !UI.$(this).is('html')) {
+                            fixedParent = UI.$(this);
+                        }
+                    });
+
+                    editor.htmleditor.data('fixedParents', false);
+
+                    if (fixedParent) {
+
+                        var transformed = [];
+
+                        fixedParent = fixedParent.parent().find(parents).each(function(){
+
+                            if (UI.$(this).css('transform') != 'none') {
+                                transformed.push(UI.$(this).data('transform-reset', {
+                                    'transform': this.style.transform,
+                                    '-webkit-transform': this.style.webkitTransform,
+                                    '-webkit-transition':this.style.webkitTransition,
+                                    'transition':this.style.transition
+                                }).css({
+                                    'transform': 'none',
+                                    '-webkit-transform': 'none',
+                                    '-webkit-transition':'none',
+                                    'transition':'none'
+                                }));
+                            }
+                        });
+
+                        editor.htmleditor.data('fixedParents', transformed);
+                    }
 
                     editor.editor.state.fullScreenRestore = {scrollTop: window.pageYOffset, scrollLeft: window.pageXOffset, width: wrap.style.width, height: wrap.style.height};
                     wrap.style.width  = '';
@@ -496,6 +534,12 @@
                     var info = editor.editor.state.fullScreenRestore;
                     wrap.style.width = info.width; wrap.style.height = info.height;
                     window.scrollTo(info.scrollLeft, info.scrollTop);
+
+                    if (editor.htmleditor.data('fixedParents')) {
+                        editor.htmleditor.data('fixedParents').forEach(function(parent){
+                            parent.css(parent.data('transform-reset'));
+                        });
+                    }
                 }
 
                 setTimeout(function() {