+++ src/js/components/slideshow-fx.js
@@ -79,7 +79,9 @@
             this.container.append(ghost);
 
             ghost.children().last().on(UI.support.transition.end, function() {
-                ghost.remove();
+
+                setTimeout(ghost.remove.bind(ghost), 0);
+
                 d.resolve();
             });
 
@@ -154,7 +156,9 @@
             ghost.width();
 
             ghost.children().first().on(UI.support.transition.end, function() {
-                ghost.remove();
+
+                setTimeout(ghost.remove.bind(ghost), 0);
+
                 d.resolve();
             }).end().css({
                 'transform': 'scaleX(1)',
@@ -227,7 +231,9 @@
                     '-webkit-transition': 'all '+$this.options.duration+'ms ease-in-out '+(50+i*25)+'ms'
                 });
             }).last().on(UI.support.transition.end, function() {
-                ghost.remove();
+
+                setTimeout(ghost.remove.bind(ghost), 0);
+
                 d.resolve();
             });
 
@@ -331,7 +337,9 @@
             }
 
             boxes.last().on(UI.support.transition.end, function() {
-                ghost.remove();
+
+                setTimeout(ghost.remove.bind(ghost), 0);
+
                 d.resolve();
             });