+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+
+### 2.26.4 (Jul 13, 2016)
+
+  - Added Htmleditor enablescripts option
+  - Removed marked lib / load from a CDN
+  - Fixed slider in rtl mode
+  - Fixed jQuery v.3 switcher + slideshow animation compatibility
+  - Fixed Htmleditor fullscreen mode in fixed container
+  - Fixed prefixer gulp task
+  - Fixed stick wrapper margin in IE
+  - Fixed for Dynamic Grid ```_getSize``` is not a function
+  - Fixed where many requestanimationframes can exist (custom scroll observer)
+  - Fixed dragging items over iframes (sortable/nestable component)
+  - Fixed sortable for Chrome/Windows mobile
+
 ### 2.26.3 (May 12, 2016)
 
   - Added UIkit.Utils.throttle helper function